#3505: PhNamed deleted
