+++ eo-runtime/src/main/java/org/eolang/PhSafe.java
@@ -25,6 +25,7 @@
 package org.eolang;
 
 import EOorg.EOeolang.EOerror;
+import java.util.function.Supplier;
 
 /**
  * It catches {@link ExFailure} and
@@ -125,8 +126,54 @@ public final class PhSafe implements Phi {
             return action.act();
         } catch (final ExFailure ex) {
             throw new EOerror.ExError(
-                new Data.ToPhi(new EOerror.ErrorMsg(ex).get())
+                new Data.ToPhi(new PhSafe.ErrorMsg(ex).get())
             );
+        }
+    }
+
+    /**
+     * Builds error message from the exception.
+     *
+     * @since 0.40
+     */
+    public static final class ErrorMsg implements Supplier<String> {
+        /**
+         * The exception.
+         */
+        private final Throwable exp;
+
+        /**
+         * Ctor.
+         *
+         * @param exp The exception.
+         */
+        public ErrorMsg(final Throwable exp) {
+            this.exp = exp;
+        }
+
+        @Override
+        public String get() {
+            return ErrorMsg.message(this.exp);
+        }
+
+        /**
+         * Make a message from an exception.
+         * @param exp The exception.
+         * @return Message.
+         */
+        private static String message(final Throwable exp) {
+            final StringBuilder ret = new StringBuilder(0);
+            if (exp.getMessage() != null) {
+                if (!(exp instanceof ExFailure)) {
+                    ret.append(exp.getClass().getSimpleName()).append(": ");
+                }
+                ret.append(exp.getMessage().replace("%", "%%"));
+            }
+            if (exp.getCause() != null) {
+                ret.append("; caused by ")
+                    .append(ErrorMsg.message(exp.getCause()));
+            }
+            return ret.toString();
         }
     }
